requirements.txt
djangorestframework
psycopg2-binary
Pillow
python-decouple
drf-yasg
django-filter
celery
redis
djoser
djangorestframework_simplejwt
django-cors-headers
python-slugify
django-allauth
django-rest-auth
django
gunicorn
coreapi
